Fixes #129: rename package, should not be breaking change since it's a internal API not public facing
