+++ vertx-jwt/src/main/java/io/vertx/ext/jwt/JWT.java
@@ -13,7 +13,7 @@
  *
  *  You may elect to redistribute this code under either of these licenses.
  */
-package io.vertx.ext.auth.jwt;
+package io.vertx.ext.jwt;
 
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
